main.py
import streamlit as st
import librosa
import numpy as np
import matplotlib.pyplot as plt

# Streamlit UI
st.title("Few Shot Language Agnostic Keyword Spotting (FSLAKWS)")

# Step 1: Audio File Upload
st.header("Step 1: Upload Audio File")
audio_file = st.file_uploader("Upload your audio file", type=["wav", "mp3", "flac"])

if audio_file is not None:
    # Load and display audio waveform
    audio_data, sr = librosa.load(audio_file, sr=None)
    st.audio(audio_file, format="audio/wav")
    
    # Plot waveform
    fig, ax = plt.subplots()
    ax.plot(np.linspace(0, len(audio_data) / sr, len(audio_data)), audio_data)
    ax.set_title("Audio Waveform")
    ax.set_xlabel("Time (s)")
    ax.set_ylabel("Amplitude")
    st.pyplot(fig)
    
    # Show spectrogram
    st.header("Audio Spectrogram")
    S = librosa.feature.melspectrogram(y=audio_data, sr=sr, n_mels=128)
    plt.figure(figsize=(10, 4))
    librosa.display.specshow(librosa.power_to_db(S, ref=np.max), y_axis='mel', x_axis='time')
    plt.title('Mel-frequency spectrogram')
    st.pyplot(plt)

# Step 2: Few-Shot Keyword Training Input
st.header("Step 2: Few Shot Keyword Training")
keywords = st.text_area("Enter Keywords to Train (comma separated)", placeholder="Keyword1, Keyword2, Keyword3")

# Collect few shot examples
if keywords:
    keyword_list = [keyword.strip() for keyword in keywords.split(",")]
    st.write("Training with the following keywords:", keyword_list)

# Step 3: Simulate Keyword Spotting (No Model Inference, Just UI)
st.header("Step 3: Simulate Keyword Spotting")
st.write("Click below to simulate keyword spotting on the uploaded audio.")

# Button to simulate keyword spotting (no model inference in this case)
if st.button("Simulate Keyword Spotting"):
    if audio_file and keywords:
        # Simulate the classification process
        predicted_keywords = {keyword: np.random.uniform(0.5, 1.0) for keyword in keyword_list}  # Dummy predictions
        
        # Display simulated results
        st.write("Simulated Predicted Keywords and Confidence Scores:")
        for keyword, score in predicted_keywords.items():
            st.write(f"{keyword}: {score:.2f}")
        
        # Performance Metrics (Simulated)
        latency = np.random.uniform(0.1, 0.5)  # Simulated latency in seconds
        throughput = len(audio_data) / latency  # Simulated throughput (samples per second)
        
        # Display simulated performance metrics
        st.write(f"Simulated Latency: {latency:.4f} seconds")
        st.write(f"Simulated Throughput: {throughput:.4f} samples/second")
        
        # Display simulated model size
        model_size = "50 MB"  # Placeholder for model size
        st.write(f"Simulated Model Size: {model_size}")
        
        # Visualization of Keyword Spotting Confidence (Simulated)
        st.subheader("Simulated Confidence Scores for Keywords")
        keyword_names = list(predicted_keywords.keys())
        scores = list(predicted_keywords.values())
        
        fig, ax = plt.subplots()
        ax.barh(keyword_names, scores, color='skyblue')
        ax.set_xlabel('Confidence Score')
        ax.set_title('Simulated Keyword Spotting Confidence Scores')
        st.pyplot(fig)

    else:
        st.warning("Please upload an audio file and enter keywords to train.")
