requirements.txt
librosa
numpy
matplotlib.pyplot